backend/requirements.txt
fastapi
python-dotenv
uvicorn
sqlmodel
python-multipart
jose
passlib
pyjwt
passlib[bcrypt]